+++ src/decorators/string/trim.mjs
@@ -1,6 +1,6 @@
-import AdjusterBase from "../../libs/AdjusterBase";
+import BaseSchema from "../../libs/BaseSchema";
 
-export default AdjusterBase.decoratorBuilder(_adjust)
+export default BaseSchema.decoratorBuilder(_fit)
 	.init(_init)
 	.features({
 		trim: _featureTrim,
@@ -34,18 +34,18 @@ function _featureTrim(params)
 }
 
 /**
- * adjust
+ * fit
  * @param {Params-String-Trim} params parameters
- * @param {Decorator-Values} values original / adjusted values
- * @returns {boolean} end adjustment
- * @throws {AdjusterError}
+ * @param {Decorator-Values} values original / fitted values
+ * @returns {boolean} ends fitting
+ * @throws {ValueSchemaError}
  */
-function _adjust(params, values)
+function _fit(params, values)
 {
 	if(!params.flag)
 	{
 		return false;
 	}
-	values.adjusted = values.adjusted.trim();
+	values.fitted = values.fitted.trim();
 	return false;
 }