+++ src/index.mjs
@@ -1,19 +1,19 @@
 import {CAUSE, STRING, NUMERIC_STRING} from "./libs/constants";
-import adjust from "./libs/adjust";
-import boolean from "./adjusters/boolean";
-import number from "./adjusters/number";
-import string from "./adjusters/string";
-import numericString from "./adjusters/numericString";
-import email from "./adjusters/email";
-import array from "./adjusters/array";
-import object from "./adjusters/object";
+import fit from "./libs/fit";
+import boolean from "./schema/boolean";
+import number from "./schema/number";
+import string from "./schema/string";
+import numericString from "./schema/numericString";
+import email from "./schema/email";
+import array from "./schema/array";
+import object from "./schema/object";
 
 export default {
 	CAUSE,
 	STRING,
 	NUMERIC_STRING,
 
-	adjust,
+	fit,
 
 	boolean,
 	number,